+++ portal-impl/src/com/liferay/portal/security/access/control/AllowedHostsAccessControlPolicy.java
@@ -12,8 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.security.ac;
+package com.liferay.portal.security.access.control;
 
+import com.liferay.portal.kernel.security.access.control.AccessControlled;
+import com.liferay.portal.kernel.security.access.control.BaseAccessControlPolicy;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.SetUtil;
 import com.liferay.portal.kernel.util.StringUtil;
@@ -21,7 +23,6 @@ import com.liferay.portal.security.auth.AccessControlContext;
 import com.liferay.portal.security.sso.SSOUtil;
 
 import java.lang.reflect.Method;
-
 import java.util.Set;
 
 import javax.servlet.http.HttpServletRequest;
