+++ portal-service/src/com/liferay/portal/kernel/security/access/control/AccessControlThreadLocal.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.ac;
+package com.liferay.portal.kernel.security.access.control;
 
 import com.liferay.portal.kernel.util.AutoResetThreadLocal;
