+++ portal-service/src/com/liferay/portal/kernel/security/access/control/AccessControlled.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.ac;
+package com.liferay.portal.kernel.security.access.control;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
